Implement FF ANN to approx action-value function
As the number of stocks comprising the environment increases, the state
space required to store values in the Q-table becomes too large to be
properly explored.

This is the first attempt at replacing the Q-table with a functional
form approximation of the action-value function iteself (e.g. using a
neural network).

The ANN used contains a single hidden layer with ReLU nonlinear activation
functions.